ESP32: Done SD Code
Basic tests indicate it's working as intended.

Temperature logs are written to SD and then send to the Raspberry Pi flask server. If the SD cannot be accessed for some reason, we fall back to using RTC memory until the ESP32 is manually restarted.

New server routes are used and data arrangement has change: data is now sent oldest to newest rather than newest to oldest; this means flask server needs to be updated (will commit soon) when using this version.